requirements.txt
pandas==2.2.1
python-dotenv==1.0.1
requests==2.31.0
Flask==3.0.2